move server-address to core
